requirements.txt
streamlit>=1.25.0
deepface>=0.0.79
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.20.0
